carga azure

+++ domain/model/src/main/java/co/com/bancolombia/model/retotecnicobancolombiaconsumedata/ConsumeData.java
@@ -2,17 +2,15 @@ package co.com.bancolombia.model.retotecnicobancolombiaconsumedata;
 import lombok.Builder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-//import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
-//@NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
 public class ConsumeData {
     String name;
 
-    private double area;
-    private Long population;
+    private long area;
+    private long population;
 }
